refactor: Inserido esquema de validação do body parametrizado
