+++ src/server/middleware/VerifySignUp.ts
@@ -3,7 +3,7 @@ import { IUser, User } from '../models/User'
 import { RequestHandler } from "express";
 import * as yup from "yup";
 
-const signUpValidation: yup.ObjectSchema<IUser> = yup.object().shape({
+export const signUpValidation: yup.ObjectSchema<IUser> = yup.object().shape({
     username: yup.string().required().min(4).max(20),
     email: yup.string().email().required(),
     password: yup.string().required().min(6).max(20),
@@ -33,20 +33,3 @@ export const checkDuplicateUsername : RequestHandler = async (req, res, next) =>
 
     return next();
 }
-
-export const signUpBodyValidator : RequestHandler = async (req, res, next) => {
-    try {
-        await signUpValidation.validate(req.body, { abortEarly: false });
-        return next();
-    } catch (err) {
-        const yupError = err as yup.ValidationError;
-
-        const errors: Record<string, string> = {};
-        yupError.inner.forEach(error => {
-            if (!error.path) return;
-            errors[error.path] = error.message;
-        });
-        return res.status(StatusCodes.BAD_REQUEST).json({ errors });
-    }
-}
-//Function above might be worth to refactor along with ./LoginValidation.ts/BodyValidator
